@@ -15,8 +15,8 @@
     "release:windows": "set NODE_ENV=production&& gulp release",
     "pretest": "yarn run lint",
     "test": "vitest",
-    "lint": "eslint --ext .js,.vue src gulpfile.js gulp-appdmg.js",
-    "lint:fix": "eslint --fix src gulpfile.js gulp-appdmg.js",
+    "lint": "eslint --ext .js,.vue src gulpfile.mjs gulp-appdmg.js",
+    "lint:fix": "eslint --fix src gulpfile.mjs gulp-appdmg.js",
     "storybook": "start-storybook -p 6006"
   },
   "window": {
@@ -48,7 +48,7 @@
   "author": "The Betaflight open source project.",
   "license": "GPL-3.0",
   "engines": {
-    "node": "16.x"
+    "node": "18.x"
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.13.0",
@@ -101,6 +101,7 @@
     "fs-extra": "^10.1.0",
     "glob": "^8.0.3",
     "gulp": "^4.0.2",
+    "gulp-cli": "^2.3.0",
     "gulp-debian": "^0.3.2",
     "gulp-json-editor": "^2.5.6",
     "gulp-less": "^5.0.0",
@@ -109,11 +110,11 @@
     "gulp-replace": "^1.1.3",
     "gulp-sourcemaps": "^3.0.0",
     "gulp-util": "^3.0.8",
-    "gulp-xml-transformer": "^5.1.1",
+    "gulp-xml-transformer": "^7.0.0",
     "gulp-yarn": "^3.0.0",
     "gulp-zip": "^5.1.0",
     "husky": "^4.3.0",
-    "nw-builder": "^3.8.6",
+    "nw-builder": "^4.2.3",
     "os": "^0.1.2",
     "postcss": "^8.4.17",
     "rollup": "^3.9.0",
@@ -132,7 +133,7 @@
     "yarn": "^1.22.19"
   },
   "optionalDependencies": {
-    "appdmg": "^0.6.4"
+    "appdmg": "^0.6.6"
   },
   "husky": {
     "hooks": {
